backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
numpy
anthropic
sse-starlette
python-dotenv
